Implement KNOSSOS-based training preview inference
As an alternative to the HDF5-based preview inferences
that are logged to tensorboard, it is now possible to specify a region
in a KNOSSOS dataset that is used as the preview input, via the new
knossos_preview_config option for the Trainer class. Periodic preview
inference results are written to .k.zip annotation files that can be
loaded with KNOSSOS and overlayed over the original data.
.k.zip files are saved in the training directory, with file names
reflecting their training step.

This feature requires https://github.com/knossos-project/knossos_utils
to be installed.